move pointer types, traits, methods from store.rs to ptr.rs
