+++ src/circuit/circuit_frame.rs
@@ -32,7 +32,8 @@ use crate::coprocessor::Coprocessor;
 use crate::eval::{lang::Lang, Frame, Witness, IO};
 use crate::hash_witness::HashWitness;
 use crate::proof::Provable;
-use crate::store::{Ptr, Store, Thunk};
+use crate::ptr::Ptr;
+use crate::store::{Store, Thunk};
 use crate::tag::{ContTag, ExprTag, Op1, Op2};
 use num_bigint::BigUint;
 use num_integer::Integer;
