+++ src/eval/mod.rs
@@ -2,8 +2,9 @@ use crate::coprocessor::Coprocessor;
 use crate::error::ReductionError;
 use crate::field::LurkField;
 use crate::hash_witness::{ConsWitness, ContWitness};
+use crate::ptr::{ContPtr, Pointer, Ptr};
 use crate::store;
-use crate::store::{ContPtr, Expression, Pointer, Ptr, Store};
+use crate::store::{Expression, Store};
 use crate::tag::ContTag;
 use crate::writer::Write;
 use lang::Lang;
