+++ src/lib.rs
@@ -15,6 +15,7 @@ pub mod light_data;
 pub mod package;
 pub mod parser;
 pub mod proof;
+pub mod ptr;
 pub mod repl;
 pub mod scalar_store;
 pub mod store;
