+++ src/repl.rs
@@ -5,8 +5,9 @@ use crate::field::LurkField;
 use crate::light_data::{Encodable, LightData, LightStore};
 use crate::package::Package;
 use crate::parser;
+use crate::ptr::{ContPtr, Pointer, Ptr};
 use crate::scalar_store::ScalarStore;
-use crate::store::{ContPtr, Expression, Pointer, Ptr, Store};
+use crate::store::{Expression, Store};
 use crate::sym::Sym;
 use crate::tag::ContTag;
 use crate::writer::Write;
