+++ src/writer.rs
@@ -1,5 +1,6 @@
 use crate::field::LurkField;
-use crate::store::{ContPtr, Continuation, Expression, Ptr, Store};
+use crate::ptr::{ContPtr, Pointer, Ptr};
+use crate::store::{Continuation, Expression, Store};
 use crate::Sym;
 use std::io;
 
@@ -14,7 +15,6 @@ pub trait Write<F: LurkField> {
 
 impl<F: LurkField> Write<F> for Ptr<F> {
     fn fmt<W: io::Write>(&self, store: &Store<F>, w: &mut W) -> io::Result<()> {
-        use crate::store::Pointer;
         if self.is_opaque() {
             // This should never fail.
             write!(w, "<Opaque ")?;
